Improve json rendering of Data in EventGrid event
It's quite common to serialize to JSON the actual data payloads
in EventGrid events. So instead of rendering those as just strings
in the output, we try to rehydrate the JSON so we can render it
as a nested object in the output. We also rearrange the properties
and skip a couple that aren't very useful (MetadataVersion and
DataVersion).