Add support for JQ when rendering JSON event data
JQ provides an awesome and very flexible query/projection language to massage JSON payloads (see https://stedolan.github.io/jq/).

By adding jq we make the rendering much nicer if the user knows what kind of payloads he's filtering for, which is typically the case given the flexibility we provide to include and exclude properties, topics and so on.

Closes #61